refactor: ProfileDisplay 분리
