+++ frontend/app/src/components/profile/index.ts
@@ -2,3 +2,4 @@ export * from './userActions'
 export * from './Profile'
 export * from './OtherProfile'
 export * from './ProfileListItem'
+export * from './ProfileDisplay'
